% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{EstParamACtH}
\alias{EstParamACtH}
\title{Estimation of chamberlet heigth parameters ImaxACtH (Initial Maximum Chamberlet Height), maxACtHinc (Maximum Chamberlet Heights Increase)
and posACtHinc (Increase in Position of Maximum Chamberlet Height)}
\usage{
EstParamACtH(approach_ACtH, print.mod, path_plot, dev_plot, dev_plot_Ct)
}
\arguments{
\item{approach_ACtH}{input which approach should be used TRUE or FALSe}

\item{print.mod}{input if the model should be printed; TRUE or FALSE}

\item{path_plot}{path where graphs should be saved; string}

\item{dev_plot}{which extension should be used to save graphs of the exponential function; string - if = NULL then graphs are not saved}

\item{dev_plot_Ct}{which extension should be used to save graphs of the Michaelis Menten function; string - if = NULL then graphs are not saved}
}
\value{
the function returns the estimated parameters ImaxACtH, maxACtHinc and posACtHinc for each specimens as a new column in the summary table
}
\description{
Fitting the chamberlet number vs. chamberlet height series of each chamber of a specimen with a polynomial function of 2nd order.
The parameters of this function are used to calculate the maximum chamberlet height and its position within a chamber. Two different
approaches can be chosen when calculating maximum chamberlet height. Either based solely on calculations or the theoretical maximum
chamberlet height and position are compared with the actual data and if they strongly deviate from each other the actual values are used.
The maximum chamberlet height is fitted by an exponential function and the extracted function parameters are the ImaxACtH and maxACtHinc.
The position of the maximum chamberlet height is fitted by an exponential function and were function parameter a assumes 1 and
function parameter b is extracted as posACtHinc.
}
\examples{


EstParamACtH(approach = TRUE, print.mod = FALSE, path_plot ="C:/Users/user/Desktop/ggplots", dev_plot ="jpg", dev_plot_Ct = "jpg")
EstParamACtH(approach = TRUE, print.mod = FALSE, path_plot ="C:/Users/user/Desktop/ggplots", dev_plot ="jpg", dev_plot_Ct = NULL)
EstParamACtH(approach = FALSE, print.mod = TRUE, path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg", dev_plot_Ct = NULL)
}
