% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{EstParamCBL}
\alias{EstParamCBL}
\title{Estimation of chamber length parameters CBI (chamber base increase) and ICBL (initial chamber base length)}
\usage{
EstParamCBL(print.mod, path_plot, dev_plot)
}
\arguments{
\item{print.mod}{input if the model should be printed; TRUE or FALSE}

\item{path_plot}{path where graphs should be saved; string}

\item{dev_plot}{giving which extension should be used; string - if = NULL then graphs are not saved}
}
\value{
the function returns the estimated parameters CBI and ICBL for each specimens as a new column in the summary table
}
\description{
Fitting the chamber number vs. chamber base length series with a power function and extracting the biometric parameters.
}
\examples{


EstParamCBL(print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot = NULL )
EstParamCBL(print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg" )

}
