% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{EstParamCL}
\alias{EstParamCL}
\title{Estimation of chamber length parameters ICL (Initial Chamber length) and CLex (chamber length expansions)}
\usage{
EstParamCL(print.mod, path_plot, dev_plot)
}
\arguments{
\item{print.mod}{input if the model should be printed; TRUE or FALSE}

\item{path_plot}{path where graphs should be saved; string}

\item{dev_plot}{giving which extension should be used; string - if = NULL then graphs are not saved}
}
\value{
the function returns the estimated parameters ICL and CLex for each specimens as a new column in the summary table
}
\description{
Fitting the chamber number vs. chamber length series with a power function and extracting the biometric parameters.
}
\examples{


EstParamCL(print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot = NULL )
EstParamCL(print.mod=TRUE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot = "jpeg")

}
