% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{EstParamMR}
\alias{EstParamMR}
\title{Estimation of chamber length parameters IMR (initial marginal radius) and MRe (marginal radius expansions)}
\usage{
EstParamMR(print.mod, path_plot, dev_plot)
}
\arguments{
\item{print.mod}{input if the model should be printed; TRUE or FALSE}

\item{path_plot}{path where graphs should be saved; string}

\item{dev_plot}{giving which extension should be used; string - if = NULL then graphs are not saved}
}
\value{
the function returns the estimated parameters IMR and MRe for each specimens as a new column in the summary table
}
\description{
Fitting a series of rotational steps (in radians) vs. marginal radius with a power function and extracting the biometric parameters.
}
\examples{


EstParamMR(print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot = NULL )
EstParamMR(print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg" )

}
