% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{EstTF}
\alias{EstTF}
\title{Estimation of morphometric characters the morphometric characters ThMR3 (total thickness at 3 mm marginal radius),
MaxMlTh(maximal mediolateral thickness), MRmax(marginal radius at MaxMlth) and F (the Flattening ratio).}
\usage{
EstTF(estMTH, para_hybrid, para_power, print.mod, path_plot, dev_plot)
}
\arguments{
\item{estMTH}{input if MlTh should be fitted or only Th; TRUE or FALSE value}

\item{para_hybrid}{vector containing start parameter for the composite function}

\item{para_power}{vector containing start parameter for the power function}

\item{print.mod}{input if the model should be printed; TRUE or FALSE}

\item{path_plot}{path where graphs should be saved; string}

\item{dev_plot}{which extension should be used to save graphs of the exponential function; string - if = NULL then graphs are not saved}
}
\value{
the function returns the list TFdf, containing a dataframes with the varaiables MR (marginal radius), Th (Thickness),
Mlth (mediolateral thickness) per specimen, and a vector with all specimen codes
}
\description{
This function fits marginal radius versus thickness with a power function and marginal radius versus mediolateral thickness
with a composite function after Eder et al (2018). It extracts the function parameters to consecutively calculate the morphometric
character THMR3 based on the power function and MaxMlTh, MRmax and F based on the composite function.
}
\examples{


EstTF(estMTH = FALSE,print.mod=FALSE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg" )
EstTF(estMTH = TRUE,print.mod=TRUE,path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg" )
EstTF(estMTH = TRUE,para_hybrid=c(b0=2e-18,b1=5,b2=-0.005,b3=500),para_power=c(a=300,b=0.05,c=-700),print.mod=TRUE, path_plot ="C:/Users/user/Desktop/ggplots",dev_plot ="jpg")

}
