% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{ParamMean}
\alias{ParamMean}
\title{Parameter Average}
\usage{
ParamMean(param)
}
\arguments{
\item{param}{name of the parameter to be processed}
}
\value{
the function returns the mean parameter for each specimens as a new column in the summary table
}
\description{
Calculates the mean of the given biometric parameter along the mesaured chamber series per specimen.
Intended to be used for the parameters backbend angle, septal angle and perimeter ratio.
}
\examples{


ParamMean()

}
