% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{ParamPRCtL}
\alias{ParamPRCtL}
\title{Estimation of the parameter PRCt (perimeter ratio of chamberlets)}
\usage{
ParamPRCtL()
}
\value{
the function returns the estimated parameter PRCt for each specimens as a new column in the summary table
}
\description{
Calculating the mean PRCtL for each specimens using the formula described in Hohenegger and Torres (2017).
It uses the lists read in by ReadInChamberlets().
}
\examples{


ParamPRCtL()

}
