% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{ReadInChamberlets}
\alias{ReadInChamberlets}
\title{ReadIn of chamberlet measurements}
\usage{
ReadInChamberlets(Inputpath, Inputsheet, v.length, from, to)
}
\arguments{
\item{Inputpath}{source path for the excel sheet}

\item{Inputsheet}{string input of sheet name}

\item{v.length}{max. number of rows allocated per specimen}

\item{from}{string input of first excel column to be read in}

\item{to}{string input of last excel column to be read in}
}
\value{
the function returns the lists CLT_A (chamberlet area), CLT_P (chamberlet perimeter), CLT_L (chamberlet length)
}
\description{
Reads in the chamberlet measurements from an excel template with multiple sheets necessary for the calculation of the
growth-invariant and growth-independent meristic characters, which are used for analysis of morphology in nummulitid genera
with subdivided chambers.
}
\examples{
system.file("extdata","input_data.xlsx", package = "WoLF")
ReadInChamberlets(Inputpath ="C:/Users/user/Desktop/Input_data.xlsx",Inputsheet = "chamberlets",v.length = 52,from="A",to="ET")

}
