% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{ReadInSpiral}
\alias{ReadInSpiral}
\title{ReadIn of Spiral Measurements}
\usage{
ReadInSpiral(Inputpath, Inputsheets, sheetlength, rad_degree)
}
\arguments{
\item{Inputpath}{source path for the excel sheet}

\item{Inputsheets}{vector containing names of existing sheets in the template; sheet names may be changed, but the
order of the sheets must remain the same Summary -> marginal radius -> chamber base length -> backbend angle ->
septal angle -> chamber area -> chamber perimeter}

\item{sheetlength}{max. number of rows to read from excel}

\item{rad_degree}{TRUE or FALSE statement if radians and degree columns are present in the marginal radius sheet}
}
\value{
the function returns the dataframes Sum (Summary table), MR (marginal radius), CBL (Chamber base length),
 BBA (Backbend angle), SA (Septal angle), CA (chamber area), CP (chamber perimeter), CL (chamber length) and PR (Perimeter ratio)
}
\description{
Reads in the spiral measurements from an excel template with multiple sheets necessary for the calculation of the
growth-invariant and growth-independent meristic characters used for analysis of morphology in nummulitid foraminifera.
}
\examples{
system.file("extdata","input_data.xlsx", package = "WoLF")
sheets<-c("summary","marginal radius","chamber base length","backbend angle","septal angle", "chamber area", "chamber perimeter")
ReadInSpiral(Inputpath = "C:/Users/user/Desktop/Input_data.xlsx",Inputsheets = sheets,sheetlength = 50,rad_degree=TRUE)

}
