% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoLF_functionsV2.R
\name{ReadInTF}
\alias{ReadInTF}
\title{ReadIn of test thickness measurements}
\usage{
ReadInTF(Inputpath, Inputsheet, from, to, nolines, nopoints)
}
\arguments{
\item{Inputpath}{source path for the excel sheet}

\item{Inputsheet}{vector containing names of existing sheets in the template; sheet names may be changed, but the
order of the sheets must remain the same Summary -> marginal radius -> chamber base length -> backbend angle ->
septal angle -> chamber area -> chamber perimeter}

\item{from}{start column in excel}

\item{to}{end column in excel}

\item{nolines}{number of lines to read in}

\item{nopoints}{number of measurements per specimen}
}
\value{
the function returns the list TFdf, containing a dataframes with the varaiables MR (marginal radius), Th (Thickness),
Mlth (mediolateral thickness) per specimen, and a vector with all specimen codes.
}
\description{
Reads in the test thickness measurements from an excel template necessary for the calculation of the
growth-invariant and growth-independent meristic characters used for analysis of morphology of planispirally
coiling foraminifera in axial sections. Follows the approach by Eder et al. (2018)
}
